Made app run in taskbar. Made entire process more user friendly.
Utilized NotifyIcon to make app run in task bar without a window. Added notifications when the config is missing or corrupt. Point user directly to the config file's location in case of missing/error. Moved config directory to %appdata% which also means it can technically run for multiple users with different entities in HA.